wip: deploying with random unique address
